feat: declare platform + .d.tsc profiles — remove all hardcoded target arrays
- Add src/compiler/profile.js: parsePlatformDecl() parses declare platform {} from .d.tsc
- Add 12 .d.tsc profile files in src/profiles/ (desktop, avr, avr-heap, avr-coop, arm, nes, spectrum, genesis, ps2, dos, wasm, wasm32)
- Parser: declare platform {} produces DeclarePlatform AST node (was Noop)
- Remove ALL hardcoded fallback arrays from codegen.js, program.js, resolve.js
- _capabilities always set (desktop default auto-injected)
- Shared<T> check now catches allocator: static (was only none)
- CLI: loadProfile() resolves .d.tsc → .json → tsc_packages/ → local path
- CLI: legacy --target/--allocator/--scheduler auto-derive capabilities
- Test runner passes --platform alongside legacy flags
- Update 11 expected.error/test files for capability-based messages
- All 8 implementation steps from spec/09b complete
- 1409 tests passing

+++ test/cases/phase11/inline/nested/expected.c
@@ -1,7 +1,7 @@
 #include "runtime.h"
 
-typedef struct { float x; float y; } V2;
-typedef struct { V2 pos; float scale; } Transform;
+typedef struct { int32_t x; int32_t y; } V2;
+typedef struct { V2 pos; int32_t scale; } Transform;
 
 int main(void) {
     TSC_INIT();
