+++ test/cases/phase11/inline/value-type/expected.c
@@ -1,11 +1,11 @@
 #include "runtime.h"
 
-typedef struct { float x; float y; } Vec2;
+typedef struct { int32_t x; int32_t y; } Vec2;
 
 int main(void) {
     TSC_INIT();
     Vec2 v = {0};
-    v.x = 1.0f;
-    v.y = 2.0f;
+    v.x = 1;
+    v.y = 2;
     return 0;
 }
